Splitting the convolution layers. (#448)
* Splitting the convoltion layers.

* Fixing the import of densenet.

* Fixed pep8
