+++ keras_contrib/layers/__init__.py
@@ -5,8 +5,9 @@ from .advanced_activations.srelu import SReLU
 from .advanced_activations.swish import Swish
 from .advanced_activations.sinerelu import SineReLU
 
-from .convolutional import CosineConv2D
-from .convolutional import SubPixelUpscaling
+from .convolutional.cosineconvolution2d import CosineConv2D
+from .convolutional.cosineconvolution2d import CosineConvolution2D
+from .convolutional.subpixelupscaling import SubPixelUpscaling
 
 from .core import CosineDense
